+++ src/selection.c
@@ -12,7 +12,7 @@ void init_selection(const View *v, SelectionInfo *info)
     block_iter_goto_offset(&info->si, info->so);
     info->swapped = false;
     if (info->so > info->eo) {
-        long o = info->so;
+        size_t o = info->so;
         info->so = info->eo;
         info->eo = o;
         info->si = v->cursor;
@@ -36,7 +36,7 @@ void init_selection(const View *v, SelectionInfo *info)
     }
 }
 
-long prepare_selection(View *v)
+size_t prepare_selection(View *v)
 {
     SelectionInfo info;
     init_selection(v, &info);
@@ -44,7 +44,7 @@ long prepare_selection(View *v)
     return info.eo - info.so;
 }
 
-char *view_get_selection(View *v, long *size)
+char *view_get_selection(View *v, size_t *size)
 {
     char *buf = NULL;
 
@@ -61,7 +61,7 @@ char *view_get_selection(View *v, long *size)
 size_t get_nr_selected_lines(const SelectionInfo *info)
 {
     BlockIter bi = info->si;
-    long pos = info->so;
+    size_t pos = info->so;
     CodePoint u = 0;
     size_t nr_lines = 1;
 
@@ -77,7 +77,7 @@ size_t get_nr_selected_lines(const SelectionInfo *info)
 size_t get_nr_selected_chars(const SelectionInfo *info)
 {
     BlockIter bi = info->si;
-    long pos = info->so;
+    size_t pos = info->so;
     CodePoint u;
     size_t nr_chars = 0;
